d_a_suspend OK
